Allow specifying a filter callback function from models
If this filter returns false, built-in filters are skipped

For convenience, the 'config' and 'validator' attributes of the Field
class are now exposed as public, since they are needed for some basic
filtering tasks such as checking if a table is joined.

It seems there's a preference for trivial getters instead of public
attributes in the rest of the code, but my religion doesn't allow me to
write trivial getters, so it's not happening.